pyproject.toml
[project]
name = "llm-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exa-py>=1.8.7",
    "polars>=1.21.0",
    "python-dotenv>=1.0.1",
]
